Update to work with sysctl
Mostly this was just changing attribute names. I didn't spend a ton of time,
but the tests pass :)

Also changed the user/group create to not specify bare ID numbers. I couldn't
see anywhere that indicated this was a requirement, so it's now just a system
account. This was done to prevent an error on CentOS 5.x where the user create
with ID 200 also created a group with ID 200 and then failed on the group
create.